consolidated data mod benchmarks
